Fixed being invited to new channels, and allowed null responses
Fixed issue where when the bot was invited into a new chathub, that hub
was not registered correctly with the ConnectedHubs list.

Added escape condition on the Say() method to correctly handle a null
message, giving Responders the ability to "return null" when they didn't
really have anything to say.